Accept and send JS style times.

+++ server/api/app/user_types.go
@@ -1592,10 +1592,10 @@ func (ut *JSONDataMetaStationFirmware) Validate() (err error) {
 // jSONDataRow user type.
 type jSONDataRow struct {
 	D        map[string]interface{} `form:"d,omitempty" json:"d,omitempty" yaml:"d,omitempty" xml:"d,omitempty"`
-	ID       *int                   `form:"id,omitempty" json:"id,omitempty" yaml:"id,omitempty" xml:"id,omitempty"`
+	ID       *int64                 `form:"id,omitempty" json:"id,omitempty" yaml:"id,omitempty" xml:"id,omitempty"`
 	Location []float64              `form:"location,omitempty" json:"location,omitempty" yaml:"location,omitempty" xml:"location,omitempty"`
 	Metas    []int                  `form:"metas,omitempty" json:"metas,omitempty" yaml:"metas,omitempty" xml:"metas,omitempty"`
-	Time     *int                   `form:"time,omitempty" json:"time,omitempty" yaml:"time,omitempty" xml:"time,omitempty"`
+	Time     *int64                 `form:"time,omitempty" json:"time,omitempty" yaml:"time,omitempty" xml:"time,omitempty"`
 }
 
 // Validate validates the jSONDataRow type instance.
@@ -1639,10 +1639,10 @@ func (ut *jSONDataRow) Publicize() *JSONDataRow {
 // JSONDataRow user type.
 type JSONDataRow struct {
 	D        map[string]interface{} `form:"d" json:"d" yaml:"d" xml:"d"`
-	ID       int                    `form:"id" json:"id" yaml:"id" xml:"id"`
+	ID       int64                  `form:"id" json:"id" yaml:"id" xml:"id"`
 	Location []float64              `form:"location" json:"location" yaml:"location" xml:"location"`
 	Metas    []int                  `form:"metas,omitempty" json:"metas,omitempty" yaml:"metas,omitempty" xml:"metas,omitempty"`
-	Time     int                    `form:"time" json:"time" yaml:"time" xml:"time"`
+	Time     int64                  `form:"time" json:"time" yaml:"time" xml:"time"`
 }
 
 // Validate validates the JSONDataRow type instance.
